+++ src/Entities/Coupons/Restriction.php
@@ -31,17 +31,23 @@ abstract class Restriction extends Resource
 
     public const TYPE_RESTRICT_TO_PLANS = 'restrict-to-plans';
 
+    public const TYPE_RESTRICT_TO_PRODUCTS = 'restrict-to-products';
+
     public const TYPE_RESTRICT_TO_SUBSCRIPTIONS = 'restrict-to-subscriptions';
 
     public const TYPE_TOTAL_REDEMPTIONS = 'total-redemptions';
 
     public const TYPE_MINIMUM_ORDER_AMOUNT = 'minimum-order-amount';
 
+    public const TYPE_PAID_BY_TIME = 'paid-by-time';
+
     protected static $supportedTypes = [
         self::TYPE_DISCOUNTS_PER_REDEMPTION,
+        self::TYPE_PAID_BY_TIME,
         self::TYPE_REDEMPTIONS_PER_CUSTOMER,
         self::TYPE_RESTRICT_TO_INVOICES,
         self::TYPE_RESTRICT_TO_PLANS,
+        self::TYPE_RESTRICT_TO_PRODUCTS,
         self::TYPE_RESTRICT_TO_SUBSCRIPTIONS,
         self::TYPE_TOTAL_REDEMPTIONS,
         self::TYPE_MINIMUM_ORDER_AMOUNT,
@@ -71,6 +77,10 @@ abstract class Restriction extends Resource
                 $restriction = new Restrictions\DiscountsPerRedemption($data);
 
                 break;
+            case self::TYPE_PAID_BY_TIME:
+                $restriction = new Restrictions\PaidByTime($data);
+
+                break;
             case self::TYPE_REDEMPTIONS_PER_CUSTOMER:
                 $restriction = new Restrictions\RedemptionsPerCustomer($data);
 
@@ -81,6 +91,10 @@ abstract class Restriction extends Resource
                 break;
             case self::TYPE_RESTRICT_TO_PLANS:
                 $restriction = new Restrictions\RestrictToPlans($data);
+
+                break;
+            case self::TYPE_RESTRICT_TO_PRODUCTS:
+                $restriction = new Restrictions\RestrictToProducts($data);
 
                 break;
             case self::TYPE_RESTRICT_TO_SUBSCRIPTIONS: